Moved the file drop listener useEffect to a separate hook.
